Add support for "query" loader config
Webpack supports defining query options for a loader as an object when
using the single-loader notation (module.loaders[].loader) which gets
JSON.stringified implicitly and passed down to the loader as a "query
string".

When HappyPack is expected to infer the loaders from webpack's config,
it has to take that query out and pass it down to the real loaders.
Prior to this patch, the query was left untouched and instead was being
passed to the *happy* loader (by webpack), which then gets confused by
those extraneous query params, preventing it from finding its plugin
instance.

Closes #26