[WIP] Add initial support for after-compile plugin
This patch is the first in the series of introducing
awesome-typescript-loader support where it sets up a proxy channel for
the "after-compile" plugin. "watch-run" doesn't really work yet and is
needed for the loader, that'd be handled in the next patch.

APIs supported during "after-compile":

- compilation.addError(String|Error) -> void
- compilation.addFileDependency(String|Array.<String>) -> void

Note on shared state: since the mapping between threads and plugins can
be 1:N, we need to make sure that a plugin executor is called at most
once for any one thread. This is only achievable by tracking some shared
state between all plugin instances (that is scoped to the ThreadPool,
really).