Use unique compilation message ids
refs GH-60

Previously, we were using the resourcePath as a unique identifier for a
compilation request. However, this stops being unique once loaders in
different stages (pre/main/post) are processing the same file, which
were causing the callbacks to get flushed for that file and eventually
failing the assertion shown in GH-60.

Now, we use a locally generated message id for each thread that requests
a compilation, guaranteeing uniqueness in that scope.