Start threadpool ASAP
Instead of waiting for the compiler to start for us to launch the
threads, we can do it as soon as the plugin is instantiated and in case
the compiler is ready before the threadpool is, we wait for it and then
proceed.

Granted, this doesn't seem to buy any time on Linux (at least on my
machine) but https://github.com/jhnns/happypack-perf-test and GH-111
indicate that it may buy time on other OSes.

Let's hope it does!