Fall back to this._compiler for the plugin list
So it seems that the loader context in webpack2 no longer exposes the
plugin list in `this.options`, and the only accessible spot I found was
in `this._compiler` which is very sad since we're trying to fight it.

This has to be revisited when webpack stops exposing
this.{_compiler,_compilation} to loaders!

Anyway, fixes GH-132