buffer process.send() calls on Windows
refs GH-166

There seems to be a difference in how process.send() behaves with fork()
between platforms since on Windows if it's called a sufficiently large
number of times (on my box that's around 6k) the process simply hangs.
This does not seem to be the case on OS X or Linux.

The fix was to never invoke process.send() until its callback yields.