run examples with different webpack configs
so the examples are getting ridiculous with the changes to webpack
config between each version and since we're trying to cover
functionality against _each_ webpack version, the example runner has
been modified to follow some conventions that we need to apply to all
examples.

The new expected structure is as follows:

|- [name]
|- [name]/dist/versions/[version]/[mode]/*
|- [name]/{lib,src}
|- [name]/versions/[version]/[mode]/webpack.config.js
|- [name]/versions/[version]/[mode]/package.json
|- [name]/test.sh

- WHERE [name] is the example name (e.g. "babel-loader")
- WHERE [version] is webpack's major version (e.g. "1")
- WHERE [mode] is either "happy" or "vanilla"
- Build artifacts are placed under `dist/`
- Source files are placed under either `lib/` or `src/`
- Dependencies are listed in the `package.json` files under `versions/`
- Configs are defined in each [version, mode] vertex directory under `versions/`

The package "@happypack/example-utils" provides some APIs to infer some
of those directories like "e.outputDir()" and "e.exampleDir()".